Support optional function score variant
This commit supports function score functions with an
optional function variant. The function is optional as it
is perfectly valid to have a function with only a weight e.g.
a function score query with multiple functions and a score mode
of "first", where the last weight function acts as some default value.

Closes opensearch-project/opensearch-java#820

Signed-off-by: Russ Cam <russcam@canva.com>