jackson `ObjectMapper`: add JDK8 & JSR310 modules
without these various core java types introduced with JDK8 (more than a
decade ago!) are not usable. we should enable them centrally so that not
everyone has to do this manually.

note that with jackson 3.x this will no longer be needed as it includes
it by default: https://github.com/FasterXML/jackson-databind/issues/982

Signed-off-by: Ralph Ursprung <Ralph.Ursprung@avaloq.com>